MongoDatabank supports scalar and array values
Added a bit of goofiness to allow MongoDatabanken to store scalar
values as well as array values.

At store time, if it's trying to store a non-object, it makes a
wrapper object with the pkey property, value property, and a
mongodatabankshim flag.

At read time, if it sees the mongodatabankshim flag, it returns the
value property instead of the whole object.