use property-value sets in Redis with keys for faster search
For search to work efficiently, we have to keep "indices" that map a property value to an object.
We do this by keeping a set in Redis databanks for each property-value pair.

With an object like {username: "evanp", name: {last: "Prodromou", first: "Evan"}, age: 43},
and a schema like {person: {pkey: "username", indices: ["name.last"]}}, when we save
we'd have a couple of new entries:

     person:evanp -> (the object)
     databank:index:person:name.last:Prodromou -> set of [person:evanp]

This should make it easier to do a search like:

     bank.search('person', {'name.last': 'Prodromou'}, ...)